feat(skill-eval): use VSS CLI markdown integration
